Rename diag.py to diagnostics.py
