+++ python/runtime/pai/train_estimator.py
@@ -17,7 +17,8 @@ import tensorflow as tf
 from runtime.model_metadata import save_model_metadata
 from runtime.pai.pai_distributed import make_estimator_distributed_runconfig
 from runtime.tensorflow import metrics
-from runtime.tensorflow.diag import init_model, load_pretrained_model_estimator
+from runtime.tensorflow.diagnostics import (init_model,
+                                            load_pretrained_model_estimator)
 from runtime.tensorflow.get_tf_version import tf_is_version2
 from runtime.tensorflow.input_fn import input_fn
 from runtime.tensorflow.train_estimator import (estimator_save,
